requirements.txt
Flask==3.0.3
requests==2.32.3
mysql-connector-python==9.0.0
sqlparse==0.5.1
python-dotenv==1.0.1
google-genai==0.3.0
